requirements-base.txt
# ================================================================
# Dependencias base para Nopolo
# ================================================================
# 
# IMPORTANTE: Instalar PyTorch PRIMERO antes de este archivo
# Ver instrucciones en requirements.txt
#
# ================================================================

# GUI
PySide6>=6.6.0

# Audio
sounddevice>=0.4.6
pydub>=0.25.1
soundfile>=0.12.1
librosa>=0.10.0
praat-parselmouth>=0.4.3
pyworld>=0.3.4
torchcrepe>=0.0.20
torchfcpe>=0.0.4  # FCPE pitch estimator: ~3-4x faster than RMVPE on CUDA

# Procesamiento
numpy>=1.24.0
scipy>=1.10.0

# TTS
edge-tts>=6.1.0
google-cloud-texttospeech>=2.16.0
azure-cognitiveservices-speech>=1.38.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
aiohttp>=3.9.0

# Utilidades
python-dotenv>=1.0.0
av>=11.0.0
tensorboardX>=2.6

# FAISS (instalar según GPU)
# GPU: pip install faiss-gpu
# CPU: pip install faiss-cpu
# Se instalará automáticamente con la versión correcta
